feat: 添加定时任务-梦幻西游转金待办通知

+++ src/entities/schedule-task/schedule-task-log.entity.ts
@@ -10,7 +10,7 @@ export class ScheduleTaskLog {
   @Column({ comment: '执行时间', name: 'execution_time', type: 'timestamp' })
   executionTime: Date;
 
-  @Column({ comment: '耗时(秒)', name: 'consuming_time', type: 'int' })
+  @Column({ comment: '耗时(秒)', name: 'consuming_time', type: 'float' })
   consumingTime: number;
 
   @Column({
